requirements/base.txt
amqp==1.4.6
anyjson==0.3.3
beautifulsoup4==4.4.0
billiard==3.3.0.20
celery==3.1.18
dj-database-url==0.3.0
Django==1.8.3
django-absoluteuri==1.1.0
django-appconf==1.0.1
django-bootstrap-form==3.2
django-celery==3.1.16
django-jsonfield==0.9.15
django-model-utils==2.3.1
django-multiupload==0.5
django-redis-cache==1.5.3  # 0.13.0 - @@@ check for backwards incompatibilities
django-reversion==1.9.1
django-user-accounts==1.2
docutils==0.12
pinax-eventlog==1.0.0
gondor==1.2.5
gunicorn==19.3.0
kaleo==1.5
kombu==3.0.26
metron==1.3.5
opbeat==3.0.4
Pillow==2.9.0
pinax-theme-bootstrap==6.1.1
pinax-types==0.9.5
psycopg2==2.6.1
python-dateutil==2.4.2
pytz==2015.4
redis==2.10.3
reportlab==3.2.0
requests==2.7.0
six==1.9.0
git+git://github.com/ddehghan/libsvg.git@e2556361f86cbc5c547a30de080d8cd07ecec574#egg=libsvg
xlrd==0.9.4
